backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pymongo==4.6.1
python-dotenv==1.0.0
scikit-learn==1.3.2
nltk==3.8.1
pytesseract==0.3.10
Pillow==10.2.0
PyJWT==2.8.0
numpy==1.26.3
google-generativeai
pypdf2
python-docx
mongomock

